Fix project-wide illogical 403/404s caused by uncast ID comparisons
Several places already had a confirmed, previously-fixed bug (see the
comments in ClientController and the Api/V1 matchmaker controllers):
comparing an Eloquent primary key (auto-cast to int) against a plain
foreign-key column with strict === silently fails on this production
host, because that column comes back as a string from PDO here even
though it doesn't locally. The result is a legitimate owner getting
403'd (or 404'd) on their own resource for no real reason.

This was the actual cause of a reported "logged into my own account,
sent an interest, clicked Accept, got 403" bug — NikahInterestController
had the exact same unguarded $interest->receiver_profile_id ===
$myProfile->id check, just never carried the (int) fix applied
elsewhere.

Applied the same (int) cast on both sides everywhere this pattern
still existed: Nikah interests (web + API), guardian messaging, safety
block/report, private file access (CNIC/payment screenshots/photos),
donations, counseling bookings, Quran live courses/subscriptions,
Quran teacher groups, posts, blog posts, wall comments, commission
self-dealing guard, and every remaining lead/batch sub-resource check
in the matchmaker controllers. Left every already-safe primary-key-vs-
primary-key comparison (e.g. $user->id === auth()->id()) untouched.

+++ app/Http/Controllers/Admin/BlogPostController.php
@@ -105,7 +105,7 @@ class BlogPostController extends Controller
     private function authorizePost(BlogPost $blogPost): void
     {
         abort_unless(
-            $blogPost->user_id === Auth::id() || Auth::user()->hasAnyRole(['admin', 'manager']),
+            (int) $blogPost->user_id === (int) Auth::id() || Auth::user()->hasAnyRole(['admin', 'manager']),
             403
         );
     }
