+++ app/Http/Controllers/Admin/CommissionController.php
@@ -81,7 +81,7 @@ class CommissionController extends Controller
     // their own earnings even if that delegation happens later.
     private function abortIfSelfDealing(CommissionLedgerEntry $entry): void
     {
-        abort_if($entry->matchmaker_id === auth()->id(), 403, 'You cannot action a commission entry that belongs to you.');
+        abort_if((int) $entry->matchmaker_id === (int) auth()->id(), 403, 'You cannot action a commission entry that belongs to you.');
     }
 
     public function approve(CommissionLedgerEntry $entry)
